market_data/requirements.txt
aiohttp==3.10.3
asgiref==3.8.1 
beautifulsoup4==4.12.3 
certifi==2024.7.4 
Django==5.1
psycopg==3.2.1 
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.8 
hiredis==3.0.0 
lxml==5.3.0
requests==2.32.3 
selenium==4.25.0 
webdriver-manager==4.0.2 